+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - Periodic celery task to queue pending tasks
+- Local folder support
 
 ## [0.27.0] 2022-08-23